extract SendBeaconBlocksByRangeRequest
