+++ beacon-chain/sync/initial-sync/BUILD.bazel
@@ -36,8 +36,6 @@ go_library(
         "//shared/rand:go_default_library",
         "//shared/timeutils:go_default_library",
         "@com_github_kevinms_leakybucket_go//:go_default_library",
-        "@com_github_libp2p_go_libp2p_core//helpers:go_default_library",
-        "@com_github_libp2p_go_libp2p_core//mux:go_default_library",
         "@com_github_libp2p_go_libp2p_core//peer:go_default_library",
         "@com_github_paulbellamy_ratecounter//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
